Update BurpExtender.java
Added command execution of a shell script to handle a system level command to send an email of those two attachments. 
which just needs this line to function as well as mail-utils to be installed. This allows for a secure email to be sent using the systems email tooling without exposing credentials in the extension code.

echo "Vuln scan report from burpSuite" | mailx  -s 'BurpSuite Report' -a $1 -a $2 -r "fromEmailaddress@domain.com" toEmailAddress@domain.com